Add an Actions module that can be used in the various other modules
Actions adds the various `get` functions that interacts with the
actual API endpoints. It also defines a new macro that __using__
modules needs to implement in order to tell Actions where to send
the requests.
