+++ lib/pco_api/actions.ex
@@ -1,0 +1,33 @@
+defmodule PcoApi.Actions do
+  defmacro __using__(_opts) do
+    quote do
+      import PcoApi.Actions
+      use HTTPoison.Base
+
+      def get, do: get("", [])
+      def get(url) when is_binary(url), do: get(url, [])
+      def get(params) when is_list(params), do: get("", params)
+      def get({:id, id}, []), do: get(id)
+      def get(url, params) do
+        case get(url, [], params: params, hackney: [basic_auth: {PcoApi.key, PcoApi.secret}]) do
+          {:ok, %HTTPoison.Response{status_code: 200, body: body}} ->
+            body["data"]
+          {:ok, %HTTPoison.Response{body: body}} ->
+            %{"errors" => [%{"detail" => detail, "title" => title}]} = body
+            raise "Request returned non-200 response. Error: #{title}: #{detail}"
+          {:error, _} ->
+            raise "PcoApi.People error"
+        end
+      end
+
+      def process_url(url), do: api_endpoint <> url
+      def process_response_body(body), do: body |> Poison.decode!
+    end
+  end
+
+  defmacro endpoint(url) do
+    quote do
+      def unquote(:api_endpoint)(), do: unquote(url)
+    end
+  end
+end
