+++ lib/pco_api/people.ex
@@ -1,2 +1,9 @@
 defmodule PcoApi.People do
+  use PcoApi.Actions
+
+  endpoint "https://api.planningcenteronline.com/people/v2/"
+
+  def me do
+    get("me", [])
+  end
 end
