+++ lib/pco_api/query/query.ex
@@ -1,5 +1,5 @@
 defmodule PcoApi.Query do
-  def where({_, _} = param), do: where([], param)
+  def where(param) when is_tuple(param), do: where([], param)
   def where(params, {attr, value}) when is_list(params) do
     add_param(params, {"where[#{attr}]", value})
   end
